requirements.txt
bcrypt==4.0.1
cffi==1.15.1
configparser==5.3.0
cryptography==39.0.0
laiarturs-ros-api==1.0.0
paramiko==3.0.0
prettytable==3.6.0
pycparser==2.21
PyNaCl==1.5.0
wcwidth==0.2.6
